rename new folder to homeV3 and remove redundant file
