+++ datahub-web-react/src/app/homeV3/HomePage.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
 import PersonalizationLoadingModal from '@app/homeV2/persona/PersonalizationLoadingModal';
-import HomePageContent from '@app/homepageV2/HomePageContent';
-import Header from '@app/homepageV2/header/Header';
-import { PageWrapper } from '@app/homepageV2/styledComponents';
+import HomePageContent from '@app/homeV3/HomePageContent';
+import Header from '@app/homeV3/header/Header';
+import { PageWrapper } from '@app/homeV3/styledComponents';
 import { SearchablePage } from '@app/searchV2/SearchablePage';
 
 export const HomePage = () => {
