--- datahub-web-react/src/app/homepageV2/HomePageHeader.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import { HeaderWrapper } from '@app/homepageV2/styledComponents';
-
-const HomePageHeader = () => {
-    return <HeaderWrapper />;
-};
-
-export default HomePageHeader;
